brainrot-detector/requirements.txt
fastapi
uvicorn[standard]
keras
pillow
numpy
tensorflow
python-multipart